perf(header): fetch admin status once via AuthContext
Move the user_profiles.is_admin lookup out of Header (which mounts on
every page/navigation) into the root-persisted AuthContext, so the
authenticated round-trip runs once per user change instead of per
Header mount. Client value is cosmetic only; server-side verifyAdmin()
remains the source of truth for authorization.

Closes #260
